Added Quiz, based on code from Michael Dawson python programming book. Added DataEditor with first funcionality, creating data for Quiz. Modified Calculator.

+++ Calculator.py
@@ -13,11 +13,11 @@ while choice1 != "0":
     if choice1 == "1":
         while choice2 != 0:
             print("To back to the prevorious menu introduce \"0\"")
-            choice2 = int(input("Introduce number: "))
+            choice2 = float(input("Introduce number: "))
             if choice2 == 0:
                 break
             number1 = choice2
-            choice2 = int(input("Introduce factor: "))
+            choice2 = float(input("Introduce factor: "))
             number2 = choice2
             choice3 = int(input("""1. Addition.
 2. Subtraction.
@@ -52,19 +52,19 @@ while choice1 != "0":
 #Used number1 as variable in loop                
                 number1 = None
                 if choice4 == 1:
-                    number1 = int(input("Introduce factor: "))
+                    number1 = float(input("Introduce factor: "))
                     result += number1
                     print("Result: ", result)
                 elif choice4 == 2:
-                    number1 = int(input("Introduce factor: "))
+                    number1 = float(input("Introduce factor: "))
                     result -= number1
                     print("Result: ", result)
                 elif choice4 == 3:
-                    number1 = int(input("Introduce factor: "))
+                    number1 = float(input("Introduce factor: "))
                     result *= number1
                     print("Result: ", result)
                 elif choice4 == 4:
-                    number1 = int(input("Introduce factor: "))
+                    number1 = float(input("Introduce factor: "))
                     result /= number1
                     print("Result: ", result)
                 elif choice4 == 0:
